requirements.txt
aiocache==0.12.3
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.6.15
click==8.2.1
coverage==7.9.2
distro==1.9.0
dnspython==2.7.0
email-validator==2.2.0
fastapi==0.115.14
fastapi-cli==0.0.7
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
jinja2==3.1.6
jiter==0.10.0
markdown-it-py==3.0.0
markupsafe==3.0.2
mdurl==0.1.2
openai==1.93.0
packaging==25.0
pillow==11.3.0
pluggy==1.6.0
polars==1.31.0
pydantic==2.11.7
pydantic-core==2.33.2
pydantic-settings==2.10.1
pygments==2.19.2
pytest==8.4.1
pytest-asyncio==1.0.0
pytest-cov==6.2.1
python-dotenv==1.1.1
python-multipart==0.0.20
pyyaml==6.0.2
rich==14.0.0
rich-toolkit==0.14.8
shellingham==1.5.4
sniffio==1.3.1
starlette==0.46.2
tqdm==4.67.1
ty==0.0.1a13
typer==0.16.0
typing-extensions==4.14.1
typing-inspection==0.4.1
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
